Refactor so DPS specific logic is in separate file and not complicated core flow
