+++ iothub_client/samples/pnp/common/pnp_dps.h
@@ -1,0 +1,20 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+#ifndef PNP_DPS_H
+#define PNP_DPS_H
+
+#include "iothub_device_client.h"
+
+#ifndef USE_PROV_MODULE
+#error "Missing cmake flag for DPS"
+#endif
+
+//
+// _PnPHelper_CreateDeviceClientHandle_ViaDps is used to create a IOTHUB_DEVICE_CLIENT_HANDLE, invoking the DPS client
+// to retrieve the needed hub information.
+//
+// Applications should NOT invoke this function directly but instead should use PnPHelper_CreateDeviceClientHandle.
+IOTHUB_DEVICE_CLIENT_HANDLE PnP_CreateDeviceClientHandle_ViaDps(const PNP_DEVICE_CONFIGURATION* pnpDeviceConfiguration);
+
+#endif /* PNP_DPS_H */
